backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-multipart>=0.0.12
python-dotenv>=1.0
anthropic>=0.40
openai>=1.54
httpx>=0.27
